+++ app/scripts/modules/amazon/src/serverGroup/details/sections/AdvancedSettingsDetailsSection.tsx
@@ -1,22 +1,24 @@
 import React from 'react';
 
-import { CollapsibleSection, ModalInjector, confirmNotManaged } from '@spinnaker/core';
+import { CollapsibleSection, ModalInjector, confirmNotManaged, noop } from '@spinnaker/core';
 
 import { IAmazonServerGroupDetailsSectionProps } from './IAmazonServerGroupDetailsSectionProps';
 
 export class AdvancedSettingsDetailsSection extends React.Component<IAmazonServerGroupDetailsSectionProps> {
   private editAdvancedSettings = (): void => {
     const { app, serverGroup } = this.props;
-    confirmNotManaged(serverGroup, app).then(() =>
-      ModalInjector.modalService.open({
-        templateUrl: require('../advancedSettings/editAsgAdvancedSettings.modal.html'),
-        controller: 'EditAsgAdvancedSettingsCtrl as ctrl',
-        resolve: {
-          application: () => app,
-          serverGroup: () => serverGroup,
-        },
-      }),
-    );
+    confirmNotManaged(serverGroup, app)
+      .then(() =>
+        ModalInjector.modalService.open({
+          templateUrl: require('../advancedSettings/editAsgAdvancedSettings.modal.html'),
+          controller: 'EditAsgAdvancedSettingsCtrl as ctrl',
+          resolve: {
+            application: () => app,
+            serverGroup: () => serverGroup,
+          },
+        }),
+      )
+      .catch(noop);
   };
 
   public render(): JSX.Element {
