+++ app/scripts/modules/amazon/src/serverGroup/details/sections/ScalingProcessesDetailsSection.tsx
@@ -1,6 +1,14 @@
 import React from 'react';
 import { IScalingProcess } from 'amazon/domain';
-import { CollapsibleSection, confirmNotManaged, HelpField, ModalInjector, timestamp, Tooltip } from '@spinnaker/core';
+import {
+  CollapsibleSection,
+  confirmNotManaged,
+  HelpField,
+  ModalInjector,
+  noop,
+  timestamp,
+  Tooltip,
+} from '@spinnaker/core';
 import { AutoScalingProcessService } from '../scalingProcesses/AutoScalingProcessService';
 import { IAmazonServerGroupDetailsSectionProps } from './IAmazonServerGroupDetailsSectionProps';
 
@@ -22,17 +30,19 @@ export class ScalingProcessesDetailsSection extends React.Component<
 
   private toggleScalingProcesses = (): void => {
     const { app, serverGroup } = this.props;
-    confirmNotManaged(serverGroup, app).then(() =>
-      ModalInjector.modalService.open({
-        templateUrl: require('../scalingProcesses/modifyScalingProcesses.html'),
-        controller: 'ModifyScalingProcessesCtrl as ctrl',
-        resolve: {
-          serverGroup: () => serverGroup,
-          application: () => app,
-          processes: () => this.state.autoScalingProcesses,
-        },
-      }),
-    );
+    confirmNotManaged(serverGroup, app)
+      .then(() =>
+        ModalInjector.modalService.open({
+          templateUrl: require('../scalingProcesses/modifyScalingProcesses.html'),
+          controller: 'ModifyScalingProcessesCtrl as ctrl',
+          resolve: {
+            serverGroup: () => serverGroup,
+            application: () => app,
+            processes: () => this.state.autoScalingProcesses,
+          },
+        }),
+      )
+      .catch(noop);
   };
 
   private getState(props: IAmazonServerGroupDetailsSectionProps): IScalingProcessesDetailsSectionState {
