+++ app/scripts/modules/amazon/src/serverGroup/details/sections/SecurityGroupsDetailsSection.tsx
@@ -2,7 +2,14 @@ import React from 'react';
 import { chain, find, sortBy } from 'lodash';
 import { UISref } from '@uirouter/react';
 
-import { CollapsibleSection, confirmNotManaged, ISecurityGroup, ModalInjector, FirewallLabels } from '@spinnaker/core';
+import {
+  CollapsibleSection,
+  confirmNotManaged,
+  ISecurityGroup,
+  ModalInjector,
+  FirewallLabels,
+  noop,
+} from '@spinnaker/core';
 
 import { IAmazonServerGroupDetailsSectionProps } from './IAmazonServerGroupDetailsSectionProps';
 
@@ -40,17 +47,19 @@ export class SecurityGroupsDetailsSection extends React.Component<
 
   private updateSecurityGroups = (): void => {
     const { app, serverGroup } = this.props;
-    confirmNotManaged(serverGroup, app).then(() =>
-      ModalInjector.modalService.open({
-        templateUrl: require('../securityGroup/editSecurityGroups.modal.html'),
-        controller: 'EditSecurityGroupsCtrl as $ctrl',
-        resolve: {
-          application: () => app,
-          serverGroup: () => serverGroup,
-          securityGroups: () => this.state.securityGroups,
-        },
-      }),
-    );
+    confirmNotManaged(serverGroup, app)
+      .then(() =>
+        ModalInjector.modalService.open({
+          templateUrl: require('../securityGroup/editSecurityGroups.modal.html'),
+          controller: 'EditSecurityGroupsCtrl as $ctrl',
+          resolve: {
+            application: () => app,
+            serverGroup: () => serverGroup,
+            securityGroups: () => this.state.securityGroups,
+          },
+        }),
+      )
+      .catch(noop);
   };
 
   public componentWillReceiveProps(nextProps: IAmazonServerGroupDetailsSectionProps): void {
